Use connection ID as key for WebTransportSessions.
WebTrannsportSessionId is not unique for all the sessions associated
with a certain server. It starts from 0 for each connection. The old
session is destoryed when a new session with the same ID is created.

As we only allow a single WebTransport session for a QUIC connection,
it's safe to use QUIC connection ID as key.